requirements.txt
scipy==1.10.1
numpy==1.24.4
matplotlib==3.7.2
opencv_python==4.1.2.30
torch==2.0.1+cu117
torchvision==0.15.2+cu117
tqdm==4.65.0
Pillow==9.5.0
labelme==3.16.7